Move widget utility functions to utils/ui/widget.js.
