+++ packages/ckeditor5-table/src/utils/common.js
@@ -7,49 +7,8 @@
  * @module table/utils/common
  */
 
-import { isWidget } from '@ckeditor/ckeditor5-widget/src/utils';
-import { createEmptyTableCell, findAncestor, updateNumericAttribute } from '../commands/utils';
+import { createEmptyTableCell, updateNumericAttribute } from '../commands/utils';
 import TableWalker from '../tablewalker';
-
-// Checks if a given view element is a table widget.
-//
-// @param {module:engine/view/element~Element} viewElement
-// @returns {Boolean}
-function isTableWidget( viewElement ) {
-	return !!viewElement.getCustomProperty( 'table' ) && isWidget( viewElement );
-}
-
-/**
- * Returns a table widget editing view element if one is selected.
- *
- * @param {module:engine/view/selection~Selection|module:engine/view/documentselection~DocumentSelection} selection
- * @returns {module:engine/view/element~Element|null}
- */
-export function getSelectedTableWidget( selection ) {
-	const viewElement = selection.getSelectedElement();
-
-	if ( viewElement && isTableWidget( viewElement ) ) {
-		return viewElement;
-	}
-
-	return null;
-}
-
-/**
- * Returns a table widget editing view element if one is among the selection's ancestors.
- *
- * @param {module:engine/view/selection~Selection|module:engine/view/documentselection~DocumentSelection} selection
- * @returns {module:engine/view/element~Element|null}
- */
-export function getTableWidgetAncestor( selection ) {
-	const parentTable = findAncestor( 'table', selection.getFirstPosition() );
-
-	if ( parentTable && isTableWidget( parentTable.parent ) ) {
-		return parentTable.parent;
-	}
-
-	return null;
-}
 
 /**
  * Returns slot info of cells that starts above and overlaps a given row.
